+++ source/Public/New-DscResourcePowerShellHelp.ps1
@@ -97,9 +97,13 @@ function New-DscResourcePowerShellHelp
                 -and ($null -ne $_.FriendlyName)
         }
 
-        $descriptionPath = Join-Path -Path $mofSchema.DirectoryName -ChildPath 'readme.md'
+        # This is a workaround for issue #42.
+        $descriptionPath = Get-ChildItem -Path $mofSchema.DirectoryName -ErrorAction 'SilentlyContinue' |
+            Where-Object -FilterScript {
+                $_.Name -like 'readme.md'
+            }
 
-        if (Test-Path -Path $descriptionPath)
+        if ($descriptionPath)
         {
             Write-Verbose -Message ($script:localizedData.GenerateHelpDocumentMessage -f $result.FriendlyName)