+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- New-DscResourcePowerShellHelp
+  - Fixed so the cmdlet is case-insensitive when it looks for the README.md
+    file in a resource source folder ([issue #42](https://github.com/dsccommunity/DscResource.DocGenerator/issues/42)).
+
 ## [0.7.0] - 2020-07-08
 
 ### Added
@@ -20,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update README.md with the correct build task name; 'Publish_GitHub_Wiki_Content'.
 - Fixed wiki generation to correctly describe embedded instances in
   parameters and made new section for each embedded instances with
-  their parameters ([issue #37](https://github.com/dsccommunity/DscResource.DocGenerator/issues/37))
+  their parameters ([issue #37](https://github.com/dsccommunity/DscResource.DocGenerator/issues/37)).
 
 ### Changed